backend/requirements.txt

fastapi
uvicorn[standard]
python-multipart

pydantic-settings
python-dotenv
# LangChain + OpenAI + Pinecone
langchain
langchain-openai
langchain-community
openai
tiktoken
pinecone-client

# Document parsing
pypdf
python-docx
unstructured  # optional; good for varied docs



